# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  apperr.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  handler.addMemberRequest:
    properties:
      role:
        enum:
        - admin
        - member
        type: string
      user_id:
        type: string
    required:
    - role
    - user_id
    type: object
  handler.authResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/store.User'
    type: object
  handler.createWorkspaceRequest:
    properties:
      avatar_url:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - name
    type: object
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.logoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handler.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handler.registerRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - name
    - password
    type: object
  handler.tokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  handler.updateProfileRequest:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - name
    type: object
  store.Session:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  store.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  store.Workspace:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  store.WorkspaceMember:
    properties:
      joined_at:
        type: string
      role:
        type: string
      user_id:
        type: string
      workspace_id:
        type: string
    type: object
  store.WorkspaceWithRole:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Project management platform API
  title: Artemis API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token
      parameters:
      - description: Logout Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.logoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      summary: Register new user
      tags:
      - auth
  /users/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Upload a new avatar image
      parameters:
      - description: Avatar image
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Upload avatar
      tags:
      - user
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the currently logged-in user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user
  /users/profile:
    patch:
      consumes:
      - application/json
      description: Update allowed profile fields
      parameters:
      - description: Update Profile Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
  /users/sessions:
    get:
      consumes:
      - application/json
      description: List all active sessions for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Session'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: List sessions
      tags:
      - user
  /users/sessions/{id}:
    delete:
      consumes:
      - application/json
      description: Revoke a specific session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Revoke session
      tags:
      - user
  /workspaces:
    get:
      consumes:
      - application/json
      description: List all workspaces user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.WorkspaceWithRole'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: List workspaces
      tags:
      - workspace
    post:
      consumes:
      - application/json
      description: Create a new workspace
      parameters:
      - description: Create Workspace Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createWorkspaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Workspace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Create workspace
      tags:
      - workspace
  /workspaces/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Delete workspace
      tags:
      - workspace
    get:
      consumes:
      - application/json
      description: Get a specific workspace by ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Workspace'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Get workspace
      tags:
      - workspace
    put:
      consumes:
      - application/json
      description: Update workspace details
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Workspace Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createWorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Workspace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Update workspace
      tags:
      - workspace
  /workspaces/{id}/members:
    get:
      consumes:
      - application/json
      description: List all members of the workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.WorkspaceMember'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: List members
      tags:
      - workspace
    post:
      consumes:
      - application/json
      description: Add a user to the workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Add Member Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.addMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Add member
      tags:
      - workspace
  /workspaces/{id}/members/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove a user from the workspace or leave
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperr.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperr.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperr.AppError'
      security:
      - BearerAuth: []
      summary: Remove member
      tags:
      - workspace
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"